+++ src/main/java/one/talon/model/AccountDashboardStatisticCampaigns.java
@@ -31,18 +31,18 @@ import java.io.IOException;
 public class AccountDashboardStatisticCampaigns {
   public static final String SERIALIZED_NAME_LIVE = "live";
   @SerializedName(SERIALIZED_NAME_LIVE)
-  private Integer live;
+  private Long live;
 
   public static final String SERIALIZED_NAME_ENDING_SOON = "endingSoon";
   @SerializedName(SERIALIZED_NAME_ENDING_SOON)
-  private Integer endingSoon;
+  private Long endingSoon;
 
   public static final String SERIALIZED_NAME_LOW_ON_BUDGET = "lowOnBudget";
   @SerializedName(SERIALIZED_NAME_LOW_ON_BUDGET)
-  private Integer lowOnBudget;
+  private Long lowOnBudget;
 
 
-  public AccountDashboardStatisticCampaigns live(Integer live) {
+  public AccountDashboardStatisticCampaigns live(Long live) {
     
     this.live = live;
     return this;
@@ -54,17 +54,17 @@ public class AccountDashboardStatisticCampaigns {
   **/
   @ApiModelProperty(required = true, value = "Number of campaigns that are active and live (across all Applications).")
 
-  public Integer getLive() {
+  public Long getLive() {
     return live;
   }
 
 
-  public void setLive(Integer live) {
+  public void setLive(Long live) {
     this.live = live;
   }
 
 
-  public AccountDashboardStatisticCampaigns endingSoon(Integer endingSoon) {
+  public AccountDashboardStatisticCampaigns endingSoon(Long endingSoon) {
     
     this.endingSoon = endingSoon;
     return this;
@@ -76,17 +76,17 @@ public class AccountDashboardStatisticCampaigns {
   **/
   @ApiModelProperty(required = true, value = "Campaigns scheduled to expire sometime in the next 7 days.")
 
-  public Integer getEndingSoon() {
+  public Long getEndingSoon() {
     return endingSoon;
   }
 
 
-  public void setEndingSoon(Integer endingSoon) {
+  public void setEndingSoon(Long endingSoon) {
     this.endingSoon = endingSoon;
   }
 
 
-  public AccountDashboardStatisticCampaigns lowOnBudget(Integer lowOnBudget) {
+  public AccountDashboardStatisticCampaigns lowOnBudget(Long lowOnBudget) {
     
     this.lowOnBudget = lowOnBudget;
     return this;
@@ -98,12 +98,12 @@ public class AccountDashboardStatisticCampaigns {
   **/
   @ApiModelProperty(required = true, value = "Campaigns with less than 10% of budget left.")
 
-  public Integer getLowOnBudget() {
+  public Long getLowOnBudget() {
     return lowOnBudget;
   }
 
 
-  public void setLowOnBudget(Integer lowOnBudget) {
+  public void setLowOnBudget(Long lowOnBudget) {
     this.lowOnBudget = lowOnBudget;
   }
